requirements-production.txt
# Production requirements for cloud deployment (Render, Railway, etc.)
# This file explicitly avoids any GPU dependencies and ensures CPU-only TensorFlow

# Core Flask dependencies
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
flask-sqlalchemy==3.0.5
flask-bcrypt==1.0.1

# Machine Learning - CPU ONLY
tensorflow-cpu==2.13.0
# Explicitly exclude GPU versions
# tensorflow-gpu  # NEVER include this
# tensorflow      # This might pull GPU version

# Image processing
pillow==10.0.1
opencv-python-headless==4.8.1.78  # Headless version for servers

# Scientific computing
numpy==1.24.3
matplotlib==3.7.2
scipy==1.11.3

# Environment and configuration
python-dotenv==1.0.0

# Email services
flask-mail==0.9.1

# AI/ML services
openai==1.3.0

# Web server
gunicorn==21.2.0

# Security
werkzeug==2.3.7

# HTTP requests
requests==2.32.3
httpx==0.27.2